Always run before/after/around validation hooks, even when not validating the object
The main reason for this is that it is very common to want to
execute code before every save and validation.  Previously,
before_save was not called when calling Model#valid?, and
before_validation was not called when calling
Model#save(:validate=>false).  Model#_before_validation was
added to meet this need, but it's a private method that users
were not supposed to use.