Make threaded connection pools not block while connections are being made
Previously, ConnectionPool#make_new was called with the pool mutex
being held, which meant that if it takes a while to establish a
connection, all other connection pool activity stops while the
connection is being made.

This refactors the connection pool so that make_new is called
without the connection pool mutex, so only the thread that needs
the new connection will block while the connection is being
made.

This also makes sure the connection pool condition variables
correctly signal on disconnection, which previously could have
led to to waiting for the pool timeout in low traffic situations
(still returning a connection, but only after the timeout).

This also fixes the threaded connection pool specs to use mutexes
appropriately around arrays/hashes used in the specs, which should
fix some occassional spec failures on JRuby.

This modifies the ConnectionPool private API, so anyone calling
private connection pool methods may need to update their code.