Fix handling of invalid :pool_class symbols
Previously you couldn't hit the error, because the require
would raise a LoadError.  It doesn't make sense to allow requiring
random files if there is no way to register the entries.  In
Sequel 5, non default pool classes need to be specified by having
the :pool_class be a class instead of a non-default symbol.